+++ crates/client/src/rpc/impls/fee_history_cache.rs
@@ -1,9 +1,13 @@
-use crate::rpc::types::{FeeHistoryEntry, MAX_FEE_HISTORY_CACHE_BLOCK_COUNT};
+use cfx_types::{Space, H256, U256};
+use cfxcore::consensus::PhantomBlock;
 use parking_lot::RwLock;
+use primitives::{transaction::SignedTransaction, BlockHeader};
 use std::{
     collections::{BTreeMap, VecDeque},
     sync::Arc,
 };
+
+pub const MAX_FEE_HISTORY_CACHE_BLOCK_COUNT: u64 = 1024;
 
 #[derive(Debug, Clone)]
 pub struct FeeHistoryCache {
@@ -15,6 +19,70 @@ impl FeeHistoryCache {
         Self {
             inner: Arc::new(RwLock::new(FeeHistoryCacheInner::new())),
         }
+    }
+
+    /*
+        This cache is used to store the FeeHistoryEntry data for the latest 1024 blocks, enabling quick queries.
+
+        Update Logic:
+        If the cached block data is outdated, clear the cache first.
+        If the cache is empty, fetch the FeeHistoryEntry data for the latest q blocks directly from the DB and cache them (q is the number of blocks to be queried this time).
+        If the cache is not empty, get the upper bound of the cache and update from that block number to the latest block.
+        If the number of blocks in the cache exceeds 1024, delete the oldest block data until the number of blocks in the cache is 1024.
+    */
+    pub fn update_to_latest_block<F>(
+        &self, latest_block: u64, query_len: u64, fetch_block_by_height: F,
+    ) -> Result<(), String>
+    where F: Fn(u64) -> Result<PhantomBlock, String> {
+        self.check_and_clear_cache(latest_block);
+
+        let start_block = if self.is_empty() {
+            if latest_block <= query_len {
+                0
+            } else {
+                latest_block - query_len + 1
+            }
+        } else {
+            self.upper_bound() + 1
+        };
+
+        for i in start_block..=latest_block {
+            let block = fetch_block_by_height(i)?;
+            self.push_back(
+                i,
+                FeeHistoryEntry::from_block(
+                    Space::Ethereum,
+                    &block.pivot_header,
+                    block.transactions.iter().map(|x| &**x),
+                ),
+            )?;
+        }
+
+        // update cache if block changes due to reorg
+        if self.lower_bound() < start_block {
+            let mut check = start_block - 1;
+            let mut curr = self
+                .get(start_block)
+                .ok_or_else(|| "fee_history_entry not found")?;
+            while check >= self.lower_bound() {
+                let item = self
+                    .get(check)
+                    .ok_or_else(|| "fee_history_entry not found")?;
+                if curr.parent_hash == item.header_hash {
+                    break;
+                }
+                let block = fetch_block_by_height(check)?;
+                curr = FeeHistoryEntry::from_block(
+                    Space::Ethereum,
+                    &block.pivot_header,
+                    block.transactions.iter().map(|x| &**x),
+                );
+                self.update(check, curr.clone());
+                check -= 1;
+            }
+        }
+
+        Ok(())
     }
 
     pub fn max_blocks(&self) -> u64 { self.inner.read().max_blocks }
@@ -134,7 +202,7 @@ impl FeeHistoryCache {
 }
 
 #[derive(Debug)]
-pub struct FeeHistoryCacheInner {
+struct FeeHistoryCacheInner {
     /// Stores the lower bound of the cache
     lower_bound: u64,
     /// Stores the upper bound of the cache
@@ -152,6 +220,75 @@ impl FeeHistoryCacheInner {
             upper_bound: 0,
             max_blocks: MAX_FEE_HISTORY_CACHE_BLOCK_COUNT,
             entries: BTreeMap::new(),
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct FeeHistoryEntry {
+    /// The base fee per gas for this block.
+    pub base_fee_per_gas: u64,
+    /// Gas used ratio this block.
+    pub gas_used_ratio: f64,
+    /// Gas used by this block.
+    pub gas_used: u64,
+    /// Gas limit by this block.
+    pub gas_limit: u64,
+    /// Hash of the block.
+    pub header_hash: H256,
+    ///
+    pub parent_hash: H256,
+    /// Approximated rewards for the configured percentiles.
+    pub rewards: Vec<u128>,
+    /// The timestamp of the block.
+    pub timestamp: u64,
+}
+
+impl FeeHistoryEntry {
+    pub fn from_block<'a, I>(
+        space: Space, pivot_header: &BlockHeader, transactions: I,
+    ) -> Self
+    where I: Clone + Iterator<Item = &'a SignedTransaction> {
+        let gas_limit: u64 = if space == Space::Native {
+            pivot_header.core_space_gas_limit().as_u64()
+        } else {
+            pivot_header.espace_gas_limit(true).as_u64()
+        };
+
+        let gas_used = transactions
+            .clone()
+            .map(|x| *x.gas_limit())
+            .reduce(|x, y| x + y)
+            .unwrap_or_default()
+            .as_u64();
+
+        let gas_used_ratio = gas_used as f64 / gas_limit as f64;
+
+        let base_fee_per_gas =
+            pivot_header.space_base_price(space).unwrap_or_default();
+
+        let mut rewards: Vec<_> = transactions
+            .map(|tx| {
+                if *tx.gas_price() < base_fee_per_gas {
+                    U256::zero()
+                } else {
+                    tx.effective_gas_price(&base_fee_per_gas)
+                }
+            })
+            .map(|x| x.as_u128())
+            .collect();
+
+        rewards.sort_unstable();
+
+        Self {
+            base_fee_per_gas: base_fee_per_gas.as_u64(),
+            gas_used_ratio,
+            gas_used,
+            gas_limit,
+            header_hash: pivot_header.hash(),
+            parent_hash: *pivot_header.parent_hash(),
+            rewards,
+            timestamp: pivot_header.timestamp(),
         }
     }
 }
