+++ crates/client/src/rpc/types/constants.rs
@@ -3,5 +3,3 @@
 /// The tx_pool will reject the tx if the gas is larger than half of the block
 /// gas limit. which is 30_000_000 before 1559, and 60_000_000 after 1559.
 pub const MAX_GAS_CALL_REQUEST: u64 = 15_000_000;
-
-pub const MAX_FEE_HISTORY_CACHE_BLOCK_COUNT: u64 = 1024;
